backend/requirements.txt
# Backend dependencies for location tracker
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
asyncpg==0.29.0
python-multipart==0.0.6
python-dateutil==2.8.2